+++ local/defaults/profiles.ps1
@@ -53,6 +53,7 @@ $DFIRWS_PROFILES = @{
             "vscode-cpp",
             "vscode-mermaid",
             "vscode-ruff",
+            "NetworkMiner",
             # release.ps1
             "Audacity",
             "Zui",